[GFC] Move GridItemSizingFunctions to its own file. https://bugs.webkit.org/show_bug.cgi?id=318282 rdar://problem/181070532
Reviewed by David Kilzer and Vitor Roriz.

Code related to the GridItemSizingFunctions live in both
TrackSizingAlgorithm and GridLayoutUtils. The former provides some
helpers to get a specific set of functions depending on the axis the
caller wants while the former provides the definiton of the struct.
We can improve the organization of the code by just moving both of these
things into the same file so that they both live next to each other.

Canonical link: https://commits.webkit.org/316222@main
