+++ Source/WebCore/layout/formattingContexts/grid/TrackSizingAlgorithm.h
@@ -26,6 +26,7 @@
 #pragma once
 
 #include "AxisConstraint.h"
+#include "GridItemSizingFunctions.h"
 #include "GridTypeAliases.h"
 #include "LayoutUnit.h"
 #include "PlacedGridItem.h"
@@ -39,20 +40,6 @@ class StyleContentAlignmentData;
 namespace Layout {
 
 class IntegrationUtils;
-
-struct GridItemSizingFunctions {
-    GridItemSizingFunctions(Function<LayoutUnit(const PlacedGridItem&, LayoutUnit oppositeAxisConstraint)> minContentContributionFunction, Function<LayoutUnit(const PlacedGridItem&, LayoutUnit oppositeAxisConstraint)> maxContentContributionFunction,
-        Function<LayoutUnit(const PlacedGridItem&, const TrackSizingFunctionsList&, LayoutUnit borderAndPadding, LayoutUnit availableSpace, LayoutUnit oppositeAxisConstraint)> usedMinimumSizeFunction)
-            : minContentContribution(WTF::move(minContentContributionFunction))
-            , maxContentContribution(WTF::move(maxContentContributionFunction))
-            , usedMinimumSize(WTF::move(usedMinimumSizeFunction))
-    {
-    }
-
-    Function<LayoutUnit(const PlacedGridItem&, LayoutUnit oppositeAxisConstraint)> minContentContribution;
-    Function<LayoutUnit(const PlacedGridItem&, LayoutUnit oppositeAxisConstraint)> maxContentContribution;
-    Function<LayoutUnit(const PlacedGridItem&, const TrackSizingFunctionsList&, LayoutUnit borderAndPadding, LayoutUnit availableSpace, LayoutUnit oppositeAxisConstraint)> usedMinimumSize;
-};
 
 struct TrackSizingItem {
     const PlacedGridItem& gridItem;
